Move most of yarprobotinterface logic to the libYARP_robotinterface library
